backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
onnxruntime>=1.18.0
opencv-python-headless>=4.10.0
python-multipart>=0.0.9
pydantic-settings>=2.4.0
apscheduler>=3.10.4
aiofiles>=24.1.0
email-validator>=2.1.0
